Manually commit the session
When autoCommit is set to TRUE, the session is committed even when an error occurs (in finally block).
In the case that the DB decides to rollback the transaction (when a constraint is violated for example) the session should not be committed.
Doing so will cause an other exception (because the transaction is aborted) and the original error will be lost.